modularize the transform code
